# Performance improvement in Extractor