When sampling stainless' performance, the `deconstruct` function is listed as one of the most -- if not *the* most -- time consuming part of the overall execution. In order to optimise it, we can re-order the pattern matching cases to put the most frequent at the top.

The order proposed here is based (for the significantly most frequent ones) on [statistics](https://gist.github.com/mantognini/959bfbc05b3e6c4b49d4a4241d57ce61) acquired when verifying all stainless benchmark under `verification/valid`. On my workstation, I could see the following effect when executing `sbt it:test`:

version  | wall clock | user time
--|--------------|--------------
master | 39:30 min | 2771 s 
patched v1 | 32:52 min | 2349 s
patched v2 | 30:49 min | 2261 s

So the results are not negligible! :-)

You can also compare the results on laraquad:
 - [before: 18 min 44 sec](http://laraquad4.epfl.ch:9000/epfl-lara/stainless/build/1789) and 
 - [after: ~16 min 58 sec~ 16 min 26 sec](http://laraquad4.epfl.ch:9000/epfl-lara/stainless/build/1792)