@mantognini 2017-06-14
I think I can. Always good to have positive peer review. :-)

However, I don't think I can publish inox to sonatype... and therefore not apply this patch to stainless without @samarion. So the question is, should I merge PRs (once I've got positive feedback from someone) anyway?