@MikaelMayer 2017-06-14
You can still push to github even if you cannot push to sonatype.

Le mercredi 14 juin 2017, Marco Antognini <notifications@github.com> a
écrit :

> I think I can. Always good to have positive peer review. :-)
>
> However, I don't think I can publish inox to sonatype... and therefore not
> apply this patch to stainless without @samarion
> <https://github.com/samarion>. So the question is, should I merge PRs
> (once I've got positive feedback from someone) anyway?
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/epfl-lara/inox/pull/28#issuecomment-308401169>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ADbyt8vyxsMaxwKvL4WjppxHvpd7eAgHks5sD8FYgaJpZM4N5m_5>
> .
>