@samarion 2017-06-14
I'll be reading and reviewing all PRs, so I can publish to sonatype when
necessary.

About this PR, thanks this is cool. We can even do it better by using a map
from class to case. I started to write a macro about this at some point but
it's hard to be general. You can write a specialized version for inox tree
case matches which should be easy. See the macros in abide for an example.

Otherwise, see with Manos how he optimized the recursive evaluator in Leon.

On Jun 14, 2017 13:57, "Mikaël Mayer" <notifications@github.com> wrote:

> You can still push to github even if you cannot push to sonatype.
>
> Le mercredi 14 juin 2017, Marco Antognini <notifications@github.com> a
> écrit :
>
> > I think I can. Always good to have positive peer review. :-)
> >
> > However, I don't think I can publish inox to sonatype... and therefore
> not
> > apply this patch to stainless without @samarion
> > <https://github.com/samarion>. So the question is, should I merge PRs
> > (once I've got positive feedback from someone) anyway?
> >
> > —
> > You are receiving this because you commented.
> > Reply to this email directly, view it on GitHub
> > <https://github.com/epfl-lara/inox/pull/28#issuecomment-308401169>, or
> mute
> > the thread
> > <https://github.com/notifications/unsubscribe-auth/
> ADbyt8vyxsMaxwKvL4WjppxHvpd7eAgHks5sD8FYgaJpZM4N5m_5>
> > .
> >
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/epfl-lara/inox/pull/28#issuecomment-308409102>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ADKzhP9QimN1dy4DuazEEY9YbXD16rqOks5sD8qpgaJpZM4N5m_5>
> .
>