@mantognini 2017-06-14
I'll keep in mind the idea of using a Map. It indeed [looks](https://github.com/epfl-lara/leon/blob/topic/grammars/src/main/scala/leon/evaluators/TableEvaluator.scala) as a good alternative. I'll try to hack something quickly with that. But I'll stay away of the macro for now as there are more important optimisation/refactoring to be made I think.

> I'll be reading and reviewing all PRs

Good to know, I'll let you merge things then! :-)