@samarion 2017-06-14
Could you also add this optimization to the Stainless deconstructors at some point? Thanks!
Also, I noticed that you didn't change the way types are deconstructed, was there a reason for this?