@mantognini 2017-06-15
> Could you also add this optimization to the Stainless deconstructors at some point? Thanks!

Sure, I'll test that and see if it helps as well.

> I noticed that you didn't change the way types are deconstructed, was there a reason for this?

Purely a gut feeling that with significantly less `case`s the bytecode might be more efficient with one less indirection. But we never know so I'll make some experiments.